ArchLinux: Update power management instructions
Remove instructions for power off, reboot, suspend and hibernate.
According to
https://wiki.archlinux.org/index.php/systemd#Power_management:

"polkit is necessary for power management as an unprivileged user. If
you are in a local systemd-logind user session and no other session is
active, the following commands will work without root privileges. If not
(for example, because another user is logged into a tty), systemd will
automatically ask you for the root password."

polkit is a dependency of NetworkManager and should be installed.